Add progress indicator when adding data to the database
* Add rich dependency
* Make loading indicator when data is added to the database

closes #9
