+++ app/data/__main__.py
@@ -1,0 +1,7 @@
+import sys
+
+from app.data.main import main
+
+
+if __name__ == '__main__':
+    main(sys.argv[1:])
