+++ app/data/dbdata.py
@@ -186,56 +186,3 @@ class DataGenerator:
                      'JOIN customer c ON o.customer_id = c.id '
                      f"WHERE o.id = {order_id}")
         return curs.fetchone()[0]
-
-
-class DataArgParser:
-    def __init__(self, args):
-        self._parser = argparse.ArgumentParser(formatter_class=RawTextHelpFormatter,
-                                               description="""Generate delivery data an dependency data.
-
-Deliveries require that there be an Order. and Orders require that there
-be Restaurants, Customers, and Drivers in the database. This program can
-produce this required data.
-
-Restaurants will be automatically created. When an Order is created, a
-random Restaurant and Customer will be chosen. A Delivery will be created
-for each Order. Drivers will be chosen randomly for each Order/Delivery.
-A Driver may end up with multiple Deliveries.
-
-All data can be deleted with the --delete-all/-d flag. All data will be
-delete from the following tables:
-    delivery, order, restaurant, owner, driver, customer, address, user
-
-examples:
-
-    python -m app.data.dbdata --drivers 4 --custs 10  --orders 10
-    python -m app.data.dbdata --delete-all""")
-
-        self._parser.add_argument('--drivers', type=int, default=0, help='number of drivers to create')
-        self._parser.add_argument('--custs', type=int, default=0, help='number of customers to create')
-        self._parser.add_argument('--orders', type=int, default=0, help='number of orders to create')
-        self._parser.add_argument('--delete-all', '-d', action='store_true', help='delete all data')
-        self._args = self._parser.parse_args(args)
-
-    def get_args(self):
-        return self._args
-
-
-def main(_args):
-    parser = DataArgParser(_args)
-    args = parser.get_args()
-    generator = DataGenerator()
-
-    if args.delete_all:
-        generator.delete_all()
-        return
-
-    generator.generate_restaurants()
-    generator.generate_drivers(args.drivers)
-    generator.generate_customers(args.custs)
-    generator.generate_orders(args.orders)
-    generator.generate_deliveries()
-
-
-if __name__ == '__main__':
-    main(sys.argv[1:])
