+++ requirements.txt
@@ -5,6 +5,8 @@ botocore==1.21.54
 certifi==2021.5.30
 cffi==1.14.6
 charset-normalizer==2.0.6
+colorama==0.4.4
+commonmark==0.9.1
 coverage==5.5
 geographiclib==1.52
 googlemaps==4.5.3
@@ -19,12 +21,14 @@ pandas==1.3.3
 pluggy==1.0.0
 py==1.10.0
 pycparser==2.20
+Pygments==2.10.0
 pyparsing==2.4.7
 pytest==6.2.5
 python-dateutil==2.8.2
 python-dotenv==0.19.0
 pytz==2021.1
 requests==2.26.0
+rich==10.12.0
 s3transfer==0.5.0
 six==1.16.0
 toml==0.10.2
